requirements.txt
astroid==3.3.5
bcrypt==4.2.1
black==24.10.0
blinker==1.9.0
boto3==1.36.6
botocore==1.36.6
cffi==1.17.1
cfgv==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
dill==0.3.9
distlib==0.3.9
Faker==32.1.0
filelock==3.16.1
flake8==7.1.1
Flask==3.1.0
Flask-Cors==5.0.0
git-filter-repo==2.45.0
greenlet==3.1.1
identify==2.6.2
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
logging==0.4.9.6
MarkupSafe==3.0.2
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
pathspec==0.12.1
pika==1.3.2
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
psycopg2-binary==2.9.11
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.10.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
s3transfer==0.11.1
six==1.16.0
SQLAlchemy==2.0.36
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.3.0
validate-docbr==1.10.0
virtualenv==20.27.1
Werkzeug==3.1.3
